Fix settings saving
