+++ src/pages/api/settings.ts
@@ -3,10 +3,10 @@ import { createProtectedHandler, NextProtectedApiHandler } from "@saleor/app-sdk
 import type { NextApiRequest, NextApiResponse } from "next";
 
 import { saleorApp } from "../../../saleor-app";
-import { providersConfig } from "../../lib/cms/config";
 import { Setting } from "../../lib/cms/utils";
 import { createClient } from "../../lib/graphql";
 import { createSettingsManager } from "../../lib/metadata";
+import { getSettings } from "../../lib/settings";
 
 export type SettingsUpdateApiRequest = Setting[];
 
@@ -19,14 +19,6 @@ export interface SettingsApiResponse {
 // const obfuscateSecret = (secret: string) => {
 //   return "*".repeat(secret.length - 4) + secret.substring(secret.length - 4);
 // };
-
-const getFieldValue = (settingsManagerValue: string | undefined, fieldName: string) => {
-  if (!settingsManagerValue && fieldName.includes("enabled")) {
-    return "false";
-  }
-
-  return settingsManagerValue ?? "";
-};
 
 const handler: NextProtectedApiHandler = async (
   req: NextApiRequest,
@@ -46,26 +38,11 @@ const handler: NextProtectedApiHandler = async (
   console.log("Settings API called", settingsManager);
 
   if (req.method === "GET") {
-    const defaultSettingNames = Object.keys(providersConfig).map((key) => `${key}.enabled`);
-    const settingsNames = Object.entries(providersConfig).reduce(
-      (prev, [providerName, config]) => [
-        ...prev,
-        ...config.tokens.map((token) => `${providerName}.${token}`),
-      ],
-      [...defaultSettingNames] as string[]
-    );
+    const settings = await getSettings(settingsManager);
 
     return res.status(200).json({
       success: true,
-      data: await Promise.all(
-        settingsNames.map(async (fieldName) => {
-          const settingsManagerValue = await settingsManager.get(fieldName);
-          return {
-            key: fieldName,
-            value: getFieldValue(settingsManagerValue, fieldName),
-          };
-        })
-      ),
+      data: settings,
     });
   } else if (req.method === "POST") {
     const settings = req.body as SettingsUpdateApiRequest;
